Rename AppCastHandler to AppCastHelper; refactor
