+++ src/NetSparkle.Tests/SemVerLikeTests.cs
@@ -35,6 +35,9 @@ namespace NetSparkleUnitTests
         [InlineData("1.0-alpha.2", "1.0-alpha.1", 1)]
         [InlineData("1.0-alpha.1", "1.0-beta.2", -1)]
         [InlineData("1.0-rc.1", "1.0-beta.2", 1)]
+        [InlineData("1", "1", 0)]
+        [InlineData("1", "2", -1)]
+        [InlineData("2", "1", 1)]
         public void CompareTest(string left, string right, int result)
         {
             Assert.Equal(
